Parent and Separate Thread execution policies
Implemented Policy classes for StateMachine to execute in the Parent
thread context and a separate thread. Only the top most HSM will execute
in a separate thread (when using a blocking event queue).
The StateMachine class itself is now split into StateMachineDef for mostly
structural aspects of the StateMachine and StateMachine for behavior.
A policy class can be composed by using the StateMachineWithPolicy mixin
class that uses any "StateType". i.e. any type that inherits from State
and implements the methods the Policy Class expects. For e.g. Both the
Separate and Parent ThreadPolicy classes expect the StateType to
implement a dispatch and execute method. So we could pass both
StateMachine and OrthogonalStateMachine types as StateTypes.
OrthogonalStateMachine comes with its own implementation of dispatch and
execute.