Reset unique id counter prior to generating each instance.
The Id generator for States and Events is pretty naive. The idea is to
create a unique id for each instantiated State and Event that are part
of s state machine. For the implementation, a static id blindly gets
incremented every time a state is needed. This works in practical
scenarios. However, during testing, when creating multiple instances of
the same type, the ids of the same States and Events are different.

The fix was to reset the id counter before instantiating another instance of
the same state machine. It's still a hack. Fix should be to reliably
generate a "random" base id that gets attached to each type statically.
This base id can then be incremented without creating id conflicts.