Modify shared_ptrs to use raw pointers or references
This was a difficult choice. The reasoning is that the API should not
force the client to use a shared/unique ptr. As the state machine itself
is not creating new states or events internally and sharing them with
the client, smart pointers at the API level were deemed restrictive.
This way, the user is free to create their state machines as raw or smart
pointers or on the stack.