Simplify Inheritance structure
* Moved TransitionTable to Transition.h
This simplifies the definition of FsmDefinition. All implementation
details related to transitions are now in Transition.h

* Merged "Def" classes with "Executors"
The attempt was to capture structural aspects of an HSM/FSM like States,
Events, Transitions, and (Parent-Child) relationships with other HSMs in
"Def" classes that inherit from HsmDefinition or FsmDefinition types.
The "execution" aspects like dispatching an Event to the appropriate
state and executing a state machine etc. was handled by an Executor.
FsmExecutor, HsmExecutor and OrthogonalHsmExecutor were the three
Executors implemented.

However, this separation broke down when adding a child HSM as it had to
be *completely* defined i.e. some Executor Hsm had to be wrapped before
adding a child state machine to a parent's Definition. This was
non-intuitive. So it made more sense to integrate this separation into
one HSMExecutor that also captured these structural aspects. As a
result, HsmExecutor became plain Hsm and OrthogonalHsmExecutor was
renamed to OrthogonalHsm. FsmExecutor was merged with HsmExecutor with
the added penalty of a couple of conditional checks and dynamic_casts.
Trade-offs.

* Explicitly deleted copy and move constructors
It does not make sense to move or copy state machines or event queues.
Hence, explicitly =delete(d) these constructors.

* Updated README.md, tests and example to reflect above changes